rename dir 'internal' to 'pkg'
commit-id:695d2bda
