+++ fakestorage/bucket.go
@@ -11,7 +11,7 @@ import (
 	"net/http"
 	"regexp"
 
-	"github.com/fsouza/fake-gcs-server/internal/backend"
+	"github.com/fsouza/fake-gcs-server/pkg/backend"
 	"github.com/gorilla/mux"
 )
