+++ fakestorage/object.go
@@ -18,8 +18,8 @@ import (
 	"time"
 
 	"cloud.google.com/go/storage"
-	"github.com/fsouza/fake-gcs-server/internal/backend"
-	"github.com/fsouza/fake-gcs-server/internal/notification"
+	"github.com/fsouza/fake-gcs-server/pkg/backend"
+	"github.com/fsouza/fake-gcs-server/pkg/notification"
 	"github.com/gorilla/mux"
 )
