+++ fakestorage/object_test.go
@@ -19,7 +19,7 @@ import (
 	"time"
 
 	"cloud.google.com/go/storage"
-	"github.com/fsouza/fake-gcs-server/internal/checksum"
+	"github.com/fsouza/fake-gcs-server/pkg/checksum"
 	"google.golang.org/api/iterator"
 )
