+++ fakestorage/server_test.go
@@ -16,8 +16,8 @@ import (
 	"testing"
 
 	"cloud.google.com/go/storage"
-	"github.com/fsouza/fake-gcs-server/internal/backend"
-	"github.com/fsouza/fake-gcs-server/internal/notification"
+	"github.com/fsouza/fake-gcs-server/pkg/backend"
+	"github.com/fsouza/fake-gcs-server/pkg/notification"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/api/iterator"
 )
