+++ pkg/backend/memory.go
@@ -12,7 +12,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/fsouza/fake-gcs-server/internal/checksum"
+	"github.com/fsouza/fake-gcs-server/pkg/checksum"
 )
 
 const timestampFormat = "2006-01-02T15:04:05.999999Z07:00"
