+++ pkg/config/config_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 
 	"github.com/fsouza/fake-gcs-server/fakestorage"
-	"github.com/fsouza/fake-gcs-server/internal/notification"
+	"github.com/fsouza/fake-gcs-server/pkg/notification"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
 )
