+++ pkg/notification/event_test.go
@@ -12,7 +12,7 @@ import (
 	"testing"
 
 	"cloud.google.com/go/pubsub"
-	"github.com/fsouza/fake-gcs-server/internal/backend"
+	"github.com/fsouza/fake-gcs-server/pkg/backend"
 )
 
 type mockPublisher struct {
